extract out TableFilterComboBox.js and flattenRecursive.js
